+++ lib/src/instrument.rs
@@ -8,33 +8,31 @@ use crate::scales;
 use crate::taper;
 use crate::wavetable::Wavetable;
 
-const DEGREES: usize = 4;
+const DEGREES: usize = 3;
 
-const CHORDS: [[i8; DEGREES]; 24] = [
-    [1, 0, 0, 0],
-    [1, 2, 0, 0],
-    [1, 3, 0, 0],
-    [1, 4, 0, 0],
-    [1, 5, 0, 0],
-    [1, 6, 0, 0],
-    [1, 7, 0, 0],
-    [1, 3, 5, 0],
-    [1, 2, 5, 0],
-    [1, 4, 5, 0],
-    [1, 5, 7, 0],
-    [1, 3, 7, 0],
-    [1, 4, 7, 0],
-    [1, 2, 7, 0],
-    [1, 5, 9, 0],
-    [1, 2, 9, 0],
-    [1, 7 + 5, 7 + 3, 0],
-    [1, 7 + 7, 7 + 3, 0],
-    [1, 7 + 9, 7 + 3, 0],
-    [1, 3, 5, 0],
-    [-4, 1, 3, 0],
-    [-6, 1, 5, 0],
-    [1, 3, 5, 7],
-    [1, 5, 7, 10],
+const CHORDS: [[i8; DEGREES]; 22] = [
+    [1, 0, 0],
+    [1, 2, 0],
+    [1, 3, 0],
+    [1, 4, 0],
+    [1, 5, 0],
+    [1, 6, 0],
+    [1, 7, 0],
+    [1, 3, 5],
+    [1, 2, 5],
+    [1, 4, 5],
+    [1, 5, 7],
+    [1, 3, 7],
+    [1, 4, 7],
+    [1, 2, 7],
+    [1, 5, 9],
+    [1, 2, 9],
+    [1, 7 + 5, 7 + 3],
+    [1, 7 + 7, 7 + 3],
+    [1, 7 + 9, 7 + 3],
+    [1, 3, 5],
+    [-4, 1, 3],
+    [-6, 1, 5],
 ];
 
 const DETUNES: [[DetuneConfig; DEGREES]; 4] = [
@@ -42,22 +40,18 @@ const DETUNES: [[DetuneConfig; DEGREES]; 4] = [
         DetuneConfig::Disabled,
         DetuneConfig::Disabled,
         DetuneConfig::Disabled,
-        DetuneConfig::Disabled,
     ],
     [
         DetuneConfig::SingleSide(0.5, 0.5 + 0.02),
         DetuneConfig::Disabled,
         DetuneConfig::Disabled,
-        DetuneConfig::Disabled,
     ],
     [
         DetuneConfig::SingleSide(0.5, 0.5 + 0.02),
         DetuneConfig::SingleSide(0.5, 0.5 + 0.02),
         DetuneConfig::SingleSide(0.5, 0.5 + 0.02),
-        DetuneConfig::SingleSide(0.5, 0.5 + 0.02),
     ],
     [
-        DetuneConfig::BothSides(1.0, 1.01),
         DetuneConfig::BothSides(1.0, 1.01),
         DetuneConfig::BothSides(1.0, 1.01),
         DetuneConfig::BothSides(1.0, 1.01),
@@ -82,7 +76,6 @@ impl<'a> Instrument<'a> {
             chord_degrees: CHORDS[0],
             amplitude: 1.0,
             degrees: [
-                Degree::new(wavetable_banks, sample_rate),
                 Degree::new(wavetable_banks, sample_rate),
                 Degree::new(wavetable_banks, sample_rate),
                 Degree::new(wavetable_banks, sample_rate),